requirements.txt
Django==3.2.4
djangorestframework==3.12.4
api==0.0.7
django-filter==2.4.0
gunicorn==20.0.4